refactor: Update Blob model to include service_url_for_direct_upload method
The Blob model has been updated to include a new method called service_url_for_direct_upload. This method returns a URL that can be used to directly upload a file for this blob on the service. The URL is intended to be short-lived for security and is generated on-demand by the client-side JavaScript responsible for uploading. This change improves the functionality and usability of the Blob model.