Add from_hash method to StorageTables::Attachable::Changes::CreateOne
This commit adds a new method called from_hash to the StorageTables::Attachable::Changes::CreateOne module. The from_hash method is responsible for handling the creation of a new blob based on the attachable data provided as a hash. It checks if the attachable data contains a :blob key or a :checksum key, and performs the appropriate action accordingly. This change improves the flexibility and usability of the CreateOne module.