Add from_hash method to StorageTables::Attachable::Changes::CreateOne (#50)
* Add from_hash method to StorageTables::Attachable::Changes::CreateOne

This commit adds a new method called from_hash to the StorageTables::Attachable::Changes::CreateOne module. The from_hash method is responsible for handling the creation of a new blob based on the attachable data provided as a hash. It checks if the attachable data contains a :blob key or a :checksum key, and performs the appropriate action accordingly. This change improves the flexibility and usability of the CreateOne module.

* refactor: Update StorageTables::Attachable::Changes::CreateOne to use find_by_checksum! method

The StorageTables::Attachable::Changes::CreateOne module has been updated to use the find_by_checksum! method instead of find_by_checksum method when retrieving a blob by its checksum. This change ensures that an exception is raised if the blob with the given checksum is not found, improving the reliability and error handling of the module.