Configurable url from rails application (#52)
* refactor: Update StorageTables::Attachable::Changes::CreateOne to use find_by_checksum! method

* Add storage tables current class to use with url_options

* Move concern to correct folder

* Add tests for disk service generating urls

* refactor: Update DiskController to handle direct file uploads

* refactor: Update DiskController to handle direct file uploads

* refactor: Update DiskService to handle missing url_options

The DiskService in the StorageTables module has been updated to handle cases where the url_options are missing. Previously, if the url_options were blank, an ArgumentError was raised. Now, a new check has been added to raise an ArgumentError with a more descriptive error message if the url_options are not set. This change improves the error handling and usability of the DiskService.