Fix module name conflicting with class name (#48)
* refactor: rename module to services

* refactor: Update module names to use 'services' instead of 'service'

* refactor: Update require statement for class_name in StorageTables::Configurator

The require statement for the class_name in the resolve method of StorageTables::Configurator has been updated to use the correct path. The previous path was using 'service' instead of 'services'. This change ensures that the correct service adapter is loaded for the given class_name.

* Delete db/migrate/20240704135631_prevent_deletion_of_blob.rb

* refactor: Update require statement for class_name in StorageTables::Configurator

* refactor: Remove unnecessary require statement in StorageTables::Service

The require statement for "storage_tables/services/configurator" has been removed from the StorageTables::Service class. This change eliminates the unnecessary dependency and improves the code cleanliness.

* refactor: Use nested classes