refactor: Update content_disposition_with method in StorageTables::Service
This commit updates the content_disposition_with method in the StorageTables::Service module. It adds support for the 'filename' and 'type' parameters, allowing the user to specify the filename and type of the content disposition. The method now checks if the type parameter is either 'attachment' or 'inline' and defaults to 'inline' if not specified. It also sanitizes the filename before formatting the content disposition using the ActionDispatch::Http::ContentDisposition.format method.

Co-authored-by: [Author Name]