Refactor base controller and disk controller (#59)
* Refactor base controller and disk controller

This commit refactors the base controller and disk controller in the StorageTables module. The base controller is now a subclass of ActionController::Base and includes the StorageTables::SetCurrent module. It also sets protect_from_forgery with :exception and disables etag_with_template_digest. The disk controller now inherits from the base controller instead of ActiveStorage::BaseController. This change improves code organization and maintainability.

Closes #XX

* Refactor .gitignore and base controller, add protect_from_forgery to disk controller

* Refactor base controller and disk controller

* Refactor base controller: Remove unnecessary line at end of file

* Remove fileserver as it is not needed