Prevent deletion of blobs with attachments (#51)
* Prevent deletion of blobs with attachments

This commit adds a migration to prevent the deletion of blobs that have attachments associated with them. It creates a trigger function `storage_tables_prevent_delete()` that raises an exception if the `attachments_count` of the blob being deleted is not zero. The migration also creates triggers on each partition table to enforce this constraint. This change improves data integrity and prevents accidental deletion of blobs with attachments.

* Add missing migration version to database structure

* Add test to prevent deletion of blob through SQL after attachment

* Fix error assertion in blob deletion test to use ActiveRecord::StatementInvalid

* Add test for mirroring files from primary to secondary services when they already exist